+++ resources/views/manage/archives.blade.php
@@ -1,0 +1,9 @@
+@extends('layouts.manage')
+
+@section('content')
+<div class="row justify-content-center">
+    <div class="col-12">
+        <manage-archives></manage-archives>
+    </div>
+</div>
+@endsection